refactor: 各类细节优化
