+++ mqttclient/include/RyanMqttUtile.h
@@ -33,19 +33,21 @@ extern void RyanMqttCleanSession(RyanMqttClient_t *client);
 extern RyanMqttError_e RyanMqttSendPacket(RyanMqttClient_t *client, uint8_t *buf, uint32_t length);
 extern RyanMqttError_e RyanMqttRecvPacket(RyanMqttClient_t *client, uint8_t *buf, uint32_t length);
 
+// msg
 extern RyanMqttBool_e RyanMqttMsgIsMatch(RyanMqttMsgHandler_t *msgHandler, const char *topic, uint16_t topicLen,
 					 RyanMqttBool_e topicMatchedFlag);
 extern RyanMqttError_e RyanMqttMsgHandlerCreate(RyanMqttClient_t *client, const char *topic, uint16_t topicLen,
 						uint16_t packetId, RyanMqttQos_e qos,
 						RyanMqttMsgHandler_t **pMsgHandler);
 extern void RyanMqttMsgHandlerDestory(RyanMqttClient_t *client, RyanMqttMsgHandler_t *msgHandler);
-extern RyanMqttError_e RyanMqttMsgHandlerFind(RyanMqttClient_t *client, const char *topic, uint16_t topicLen,
+extern RyanMqttError_e RyanMqttMsgHandlerFind(RyanMqttClient_t *client, RyanMqttMsgHandler_t *findMsgHandler,
 					      RyanMqttBool_e topicMatchedFlag, RyanMqttMsgHandler_t **pMsgHandler);
-extern void RyanMqttMsgHandlerFindByPackId(RyanMqttClient_t *client, const char *topic, uint16_t topicLen,
-					   uint16_t packetId, RyanMqttBool_e isPacketIdEqual);
+extern void RyanMqttMsgHandlerFindAndDestoryByPackId(RyanMqttClient_t *client, RyanMqttMsgHandler_t *findMsgHandler,
+						     RyanMqttBool_e isSkipMatchingId);
 extern RyanMqttError_e RyanMqttMsgHandlerAddToMsgList(RyanMqttClient_t *client, RyanMqttMsgHandler_t *msgHandler);
 extern RyanMqttError_e RyanMqttMsgHandlerRemoveToMsgList(RyanMqttClient_t *client, RyanMqttMsgHandler_t *msgHandler);
 
+// ack
 extern RyanMqttError_e RyanMqttAckHandlerCreate(RyanMqttClient_t *client, uint8_t packetType, uint16_t packetId,
 						uint16_t packetLen, uint8_t *packet, RyanMqttMsgHandler_t *msgHandler,
 						RyanMqttAckHandler_t **pAckHandler,
