+++ xmake.lua
@@ -8,9 +8,9 @@ target("RyanMqtt",function()
     set_languages("gnu99") -- 关键！启用 GNU 扩展
     set_warnings("everything") -- 启用全部警告 -Wall -Wextra -Weffc++ / -Weverything
 
-    -- set_optimize("smallest") -- -Os
+    set_optimize("smallest") -- -Os
     -- set_optimize("faster") -- -O2
-    set_optimize("fastest") -- -O3
+    -- set_optimize("fastest") -- -O3
 
     add_defines("PKG_USING_RYANMQTT_IS_ENABLE_ASSERT") -- 开启assert
 
@@ -24,12 +24,10 @@ target("RyanMqtt",function()
                 -- "-Wno-c23-extensions",
                 {force=true})
     
-
-
     --加入代码和头文件
     add_includedirs('./common', {public = true})
     add_includedirs('./coreMqtt', {public = true})
-    add_includedirs('./mqttclient', {public = true})
+    add_includedirs('./mqttclient/include', {public = true})
     add_includedirs('./platform/linux', {public = true})
     add_includedirs('./platform/linux/valloc', {public = true})
